XbrlModel: relationship/groupContent order is Decimal, serialized as a number not a string
The legacy loader emitted a network relationship's order as Arelle's r.order (a float, default
1.0). XbrlRelationship.order was declared Optional[int] and saveableValue had no float case, so
the value fell through to str(val) and serialized as "1.0" -- a string where a number belongs
(1037 such in the paclife compiled model).

- XbrlRelationship.order and XbrlGroupContent.order are now Optional[Decimal], matching the spec
  (both were int/float). XbrlLayout.order stays a str (it is an "ascending"/"descending" enum).
- saveableValue now handles float (a whole value as an int, else the float) so a numeric value is
  never stringified; the Decimal branch already emitted a number.
- json.dump gets a default= handler (_jsonDefault) as defence-in-depth: json has no native Decimal
  emitter, so any Decimal that bypasses saveableValue still serializes as a number (whole -> int),
  with an is_finite guard so Decimal('Infinity') does not raise.

Verified: paclife/principal/loreal network + groupTree relationship orders now serialize as JSON
ints (was string "1.0" for networks); no load/validation regressions.

+++ arelle/plugin/XbrlModel/XbrlGroup.py
@@ -3,6 +3,7 @@ See COPYRIGHT.md for copyright information.
 """
 
 from typing import Optional
+from decimal import Decimal
 
 from arelle.ModelValue import QName, AnyURI
 from ordered_set import OrderedSet
@@ -27,7 +28,7 @@ class XbrlGroupContent(XbrlModelObject):
     module: XbrlModuleAlias
     groupName: QName # (required) The QName of the group object that this content is included in.
     forObject: QName # (required) A single network object, cube object, domain network object or table template object QName that is included in the group object.
-    order: Optional[float] # (optional) The order of group content objects with the same groupName in the groupContents set. Default 0; ties are ordered at the processor's discretion.
+    order: Optional[Decimal] # (optional) A decimal that orders group content objects with the same groupName in the groupContents set. Default 0; ties are ordered at the processor's discretion.
 
 class XbrlGroupTree(XbrlReferencableModelObject):
     """ Group Tree Object
